requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Django==2.1.3
django-ckeditor==5.6.1
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.17.6
django-js-asset==1.1.0
django-nine==0.1.13
django-rest-elasticsearch==0.4.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
jsonschema==2.6.0
paramiko==2.4.2
psycopg2==2.8
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.0
PyNaCl==1.3.0
pytest-runner==4.4
python-dateutil==2.8.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
six==1.12.0
texttable==0.9.1
urllib3==1.24.1
websocket-client==0.56.0
